fs/mime: rename to fs/media/type
Completes the deferred rename from fs/todo/group-fs-subdirectories-by-concern.md:
media-type detection moves under fs/media/ alongside the media formats it
recognizes, freeing up the collision with the "mime" name that blocked the
bucket. Updates the sole external consumer (fs/cas/mcp/module.f.ts) and all
relative imports and doc cross-references across the tree.
