+++ fs/bnf/todo/recognizer-backend.md
@@ -16,7 +16,7 @@ Separately, several callers need to **recognize** input without building an
 AST — they only want "did it match, and what is the final state":
 
 - `fs/cas/mcp` `cas_get` metadata detection (magic-byte MIME + UTF-8 validity)
-  over a streaming blob — shipped in `fs/mime` `detectStream` with the
+  over a streaming blob — shipped in `fs/media/type` `detectStream` with the
   `A_magic`/`A_utf8` factors hand-rolled, ready to lower onto this backend;
 - "is this valid JSON / valid identifier" checks that should not allocate a tree;
 - the scanner/lexer tier of the layered parser
@@ -152,7 +152,7 @@ There is no code-point coupling in BNF itself. The string-literal constructors
 helpers** sitting *above* the agnostic core; a parallel family of **binary
 helpers** — byte / hex literals, byte sequences, byte-range sets — would author
 byte grammars the same way, all bottoming out in the agnostic `rangeEncode` /
-`oneEncode` primitives. (`fs/mime`'s `fromSentinel` hex-signature notation,
+`oneEncode` primitives. (`fs/media/type`'s `fromSentinel` hex-signature notation,
 `0x1_89_50_4e_47…n`, is a precedent for compact byte-sequence literals — just
 targeting `Vec` today rather than `RuleSet` terminals. The matcher's
 `CodePoint[]` typing is likewise nominal — structurally numbers.)
@@ -227,7 +227,7 @@ Bigger automata are built from BNF pieces in two complementary ways:
 - [new-pl.md](../../../todo/new-pl.md) — `Function.getAst` / `fromAst` (functions
   as serializable IR); `toData` is the grammar-specific case, and the serialized
   forms become content-addressable via `Object.id`
-- `fs/mime` `detectStream` — first concrete consumer (streaming MIME/UTF-8
+- `fs/media/type` `detectStream` — first concrete consumer (streaming MIME/UTF-8
   recognizer), shipped with hand-rolled `A_magic`/`A_utf8` factors that this
   backend should later replace
 - `fs/fsm`, `fs/types/byte_set`, `fs/types/range_map` — engines to reuse as the
