+++ fs/cas/mcp/README.md
@@ -117,7 +117,7 @@ such as tests.
 ### Metadata is size-independent (the default `content: false`)
 
 The metadata-only call **never buffers the blob**. It folds the CAS read stream
-through [`fs/mime`](../../mime/module.f.ts) `detectStream` — a byte-accepting
+through [`fs/media/type`](../../media/type/module.f.ts) `detectStream` — a byte-accepting
 state machine (running byte count × magic-byte signature eliminator × UTF-8
 validity DFA) that derives `{ length, mimeType, type }` in O(1) space. The
 detector stops decoding once the verdict is fixed — a magic match settles it
@@ -136,7 +136,7 @@ valid UTF-8 until a trailing invalid byte is correctly classified as `base64`
 ### Content encoding (when `content: true`)
 
 Only the `content: true` path materializes the bytes (bounded by `maxLength`). It
-classifies them with the **same** detector — [`fs/mime`](../../mime/module.f.ts)
+classifies them with the **same** detector — [`fs/media/type`](../../media/type/module.f.ts)
 `detectVec`, the single-`Vec` form of the `detectStream` machine above — so the
 three-way verdict is computed in exactly one place, never re-derived from a
 parallel `detect` + UTF-8 check. The `type` then selects whether the inline
