+++ fs/media/type/module.f.ts
@@ -32,15 +32,15 @@
  *
  * @module
  */
-import { msb, fromSentinel, length, u8List, type Vec } from '../types/bit_vec/module.f.ts'
-import { iterable } from '../types/list/module.f.ts'
-import type { Nullable } from '../types/nullable/module.f.ts'
-import { pure, type Effect, type Operation } from '../effects/module.f.ts'
-import type { List } from '../effects/list/module.f.ts'
-import type { IoResult } from '../effects/node/module.f.ts'
-import { ok, error } from '../types/result/module.f.ts'
-import { isValidCodePoint, isTextCodePoint } from '../text/code_point/module.f.ts'
-import { utf8ByteToCodePointOp, type Utf8State } from '../text/utf8/module.f.ts'
+import { msb, fromSentinel, length, u8List, type Vec } from '../../types/bit_vec/module.f.ts'
+import { iterable } from '../../types/list/module.f.ts'
+import type { Nullable } from '../../types/nullable/module.f.ts'
+import { pure, type Effect, type Operation } from '../../effects/module.f.ts'
+import type { List } from '../../effects/list/module.f.ts'
+import type { IoResult } from '../../effects/node/module.f.ts'
+import { ok, error } from '../../types/result/module.f.ts'
+import { isValidCodePoint, isTextCodePoint } from '../../text/code_point/module.f.ts'
+import { utf8ByteToCodePointOp, type Utf8State } from '../../text/utf8/module.f.ts'
 
 const { startsWith, removeFront } = msb
