+++ fs/media/type/todo/detect-json.md
@@ -2,14 +2,14 @@
 
 **Priority:** P3
 **Status:** blocked
-**Blocked by:** [fs/media/json streaming-recognizer](../../media/json/todo/streaming-recognizer.md)
+**Blocked by:** [fs/media/json streaming-recognizer](../../json/todo/streaming-recognizer.md)
 
 ### Problem
 
 The MCP server classifies stored content by content-sniffing, not by any
-stored type: `cas_get` folds the read stream through the `fs/mime` detector
+stored type: `cas_get` folds the read stream through the `fs/media/type` detector
 (`detectStream`) and reports `{ length, mime_type, type }`. The detector's
-`finish` (`fs/mime/module.f.ts:264-272`) produces a three-way verdict:
+`finish` (`fs/media/type/module.f.ts:264-272`) produces a three-way verdict:
 
 1. magic-byte hit (PNG/JPEG/GIF/WebP/PDF/ZIP) → `base64` + the detected mime;
 2. whole-blob-valid UTF-8 text → `text` + `text/plain`;
@@ -22,16 +22,16 @@ tell a JSON document from arbitrary prose. The detector should recognize
 well-formed JSON and report `application/json` (RFC 8259 / RFC 6838; UTF-8 is
 the assumed charset, so no `charset` parameter is emitted).
 
-Because the classifier is shared, fixing it in `fs/mime` fixes it everywhere:
+Because the classifier is shared, fixing it in `fs/media/type` fixes it everywhere:
 `cas_get` (`fs/cas/mcp/module.f.ts:204`) picks up `application/json`
 automatically for both the metadata-only and `content: true` paths, and any
-future `fs/mime` consumer inherits it.
+future `fs/media/type` consumer inherits it.
 
 ### Proposal
 
 Add JSON as a **refinement of the text branch**, keeping the single-classifier
 design (one machine, read off at EOF — no second, divergent copy of the rules)
-that the module documents at `fs/mime/module.f.ts:96-105`.
+that the module documents at `fs/media/type/module.f.ts:96-105`.
 
 #### 1. A fourth fold factor: a streaming JSON recognizer
 
@@ -39,7 +39,7 @@ The detector state (`DetectState`, `:201-205`) is a product of independent
 factors — bit `length` × `MagicState` × `Utf8Detect` — that meet only in
 `finish`. Add a fourth factor `A_json`: a streaming JSON **recognizer**
 (accept/reject only, no value construction) driven by the code points the UTF-8
-factor already decodes. Its core is the `fs/media/json` recognizer (§2); `fs/mime`
+factor already decodes. Its core is the `fs/media/json` recognizer (§2); `fs/media/type`
 wraps it with a one-code-point tag recording the top-level value's kind, so the
 object/array-only policy (§4) is applied at EOF — the recognizer stays pure
 (accepts any valid JSON), the MIME policy lives here:
@@ -70,7 +70,7 @@ factor its own `utf8ByteToCodePointOp` decode — at the cost of decoding twice.
 `A_json` is exactly the *"is this stream valid JSON?"* question, and it must be
 answered without buffering — otherwise the size-independence `detectStream` is
 built for is lost. Reusing `fs/media/json`'s `tokenize`/`parse` as-is does **not**
-work for two reasons that are `fs/media/json`'s to own, not `fs/mime`'s to patch:
+work for two reasons that are `fs/media/json`'s to own, not `fs/media/type`'s to patch:
 
 - `parse` builds the whole value in `top`/`stack` — O(n) memory in the document
   size.
@@ -82,14 +82,14 @@ work for two reasons that are `fs/media/json`'s to own, not `fs/mime`'s to patch
 Both are addressed by the payload-free, O(depth) recognizer proposed in
 **`fs/media/json/todo/streaming-recognizer.md`** (`recognizerInit` / `recognizerStep`
 / `recognizerAccepts`, sharing the grammar with `parse` so they cannot diverge,
-with an optional max-depth cap `fs/mime` should enable as a DoS guard). `A_json`
+with an optional max-depth cap `fs/media/type` should enable as a DoS guard). `A_json`
 is the thin §1 wrapper over it — the recognizer plus the one-code-point
 top-level tag — adding no JSON grammar of its own. This todo therefore **depends
 on** that recognizer landing first.
 
 Strictness note: the recognizer must reject raw U+0000–U+001F inside strings,
 already fixed in the shared `fs/js` tokenizer (`parseStringStateOp`). This
-matters here because `fs/mime`'s text gate admits TAB/VT/FF as text
+matters here because `fs/media/type`'s text gate admits TAB/VT/FF as text
 (`utf8Step`/`isTextCodePoint`), so without the strict check a blob like
 `{"a":"⟨TAB⟩"}` — invalid JSON per RFC 8259 — would be mislabeled
 `application/json`. `A_json` inherits the correct verdict from the recognizer
@@ -157,13 +157,13 @@ exactly the path `cas_get` uses.
       in `push`.
 - [ ] Refine `finish` to emit `application/json` for whole-blob-valid UTF-8 that
       is valid JSON **with an object/array top level** (§4 decision).
-- [ ] Add `fs/mime/proof.f.ts` cases: `{"a":1}` and `[1,2,3]` (incl. split
+- [ ] Add `fs/media/type/proof.f.ts` cases: `{"a":1}` and `[1,2,3]` (incl. split
       across chunks) → `application/json`/`text`; trailing garbage after valid
       JSON and truncated JSON → `text/plain`; non-JSON prose → `text/plain`;
       a raw TAB inside a string (`{"a":"⟨TAB⟩"}`) → `text/plain`, not
       `application/json`; bare scalars (`42`, `null`, `"hi"`, `true`) →
       `text/plain` (top-level object/array rule).
-- [ ] Update `fs/mime/module.f.ts` module doc (recognised-types table) and the
+- [ ] Update `fs/media/type/module.f.ts` module doc (recognised-types table) and the
       `cas_get` output section in `fs/cas/mcp/module.f.ts` to list
       `application/json`.
 - [ ] `npx tsc` clean; `fjs t` green with both branches of the JSON verdict
@@ -171,10 +171,10 @@ exactly the path `cas_get` uses.
 
 ### Related
 
-- `fs/mime/module.f.ts:264-272` — `finish`, where the text→JSON refinement lands.
-- `fs/mime/module.f.ts:180-195` — the UTF-8 factor whose decoded code points feed the JSON factor.
+- `fs/media/type/module.f.ts:264-272` — `finish`, where the text→JSON refinement lands.
+- `fs/media/type/module.f.ts:180-195` — the UTF-8 factor whose decoded code points feed the JSON factor.
 - `fs/media/json/todo/streaming-recognizer.md` — **blocks this**; the payload-free, O(depth) validity recognizer `A_json` wraps.
 - `fs/js/tokenizer/module.f.ts` — `parseStringStateOp`; already rejects raw U+0000–U+001F inside strings, so `A_json` inherits the correct verdict without re-deriving it.
 - `fs/media/json/parser/module.f.ts:205-238` — `foldOp` / `parse`, the grammar the recognizer reuses value-free.
 - `fs/cas/mcp/module.f.ts:196-204` — `cas_get`, the consumer that gains `application/json` for free.
-- `fs/mime/todo/single-signature-table.md` — the sibling "one source of truth" cleanup; same single-classifier principle.
+- `fs/media/type/todo/single-signature-table.md` — the sibling "one source of truth" cleanup; same single-classifier principle.
